+++ projects/clr/hipamd/src/hip_api_trace.cpp
@@ -2103,7 +2103,7 @@ HIP_ENFORCE_ABI(HipDispatchTable, hipLibraryGetManaged_fn, 503);
 //  HIP_ENFORCE_ABI_VERSIONING(<table>, 9) <- 8 + 1 = 9
 HIP_ENFORCE_ABI_VERSIONING(HipDispatchTable, 504)
 
-static_assert(HIP_RUNTIME_API_TABLE_MAJOR_VERSION == 0 && HIP_RUNTIME_API_TABLE_STEP_VERSION == 16,
+static_assert(HIP_RUNTIME_API_TABLE_MAJOR_VERSION == 0 && HIP_RUNTIME_API_TABLE_STEP_VERSION == 17,
               "If you get this error, add new HIP_ENFORCE_ABI(...) code for the new function "
               "pointers and then update this check so it is true");
 #endif