+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/api_args.h
@@ -3331,6 +3331,22 @@ typedef union rocprofiler_hip_api_args_t
         hipStream_t src;
     } hipStreamCopyAttributes;
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 17
+    struct
+    {
+        void**       dptr;
+        size_t*      bytes;
+        hipLibrary_t library;
+        const char*  name;
+    } hipLibraryGetGlobal;
+    struct
+    {
+        void**       dptr;
+        size_t*      bytes;
+        hipLibrary_t library;
+        const char*  name;
+    } hipLibraryGetManaged;
+#endif
 } rocprofiler_hip_api_args_t;
 
 ROCPROFILER_EXTERN_C_FINI