+++ projects/rocprofiler-sdk/source/include/rocprofiler-sdk/hip/runtime_api_id.h
@@ -562,5 +562,9 @@ typedef enum rocprofiler_hip_runtime_api_id_t  // NOLINT(performance-enum-size)
 #if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 16
     ROCPROFILER_HIP_RUNTIME_API_ID_hipStreamCopyAttributes,
 #endif
+#if HIP_RUNTIME_API_TABLE_STEP_VERSION >= 17
+    ROCPROFILER_HIP_RUNTIME_API_ID_hipLibraryGetGlobal,
+    ROCPROFILER_HIP_RUNTIME_API_ID_hipLibraryGetManaged,
+#endif
     ROCPROFILER_HIP_RUNTIME_API_ID_LAST,
 } rocprofiler_hip_runtime_api_id_t;